requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bikram==2.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
django-js-asset==2.0.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.4
et-xmlfile==1.1.0
Faker==16.6.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
model-bakery==1.10.1
multidict==6.0.4
numpy==1.24.1
openai==0.26.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stripe==6.1.0
tqdm==4.64.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
yapf==0.32.0
yarl==1.8.2
